src/cdstotile/requirements.txt
numpy
netcdf4
cdsapi
xarray
dask
python-dateutil
parsl
